admin/requirements.txt
# Admin UI dependencies — compatible with Python 3.9
streamlit>=1.35.0
supabase>=2.9.0
anthropic>=0.34.0
pypdf>=4.3.0
python-dotenv>=1.0.0
